refactor(app): dedupe git url exports confusing nitro auto-imports
