+++ apps/application/server/api/failure-clusters/[id]/commits.get.ts
@@ -1,7 +1,7 @@
 import { failureClusters, testRuns } from '../../../database/schema';
 import { eq } from 'drizzle-orm';
 import { optionalIntQuery } from '../../../utils/query-params';
-import { normalizeGitUrl } from '../../../utils/regression-context';
+import { normalizeGitUrl } from '../../../utils/scm/git-url';
 import { createScmProvider } from '../../../utils/scm';
 import { requireResolvedProjectAccess, requireRouteId, resolveClusterProjectId } from '../../../utils/project-access';
