+++ apps/application/server/utils/ai-context.ts
@@ -15,7 +15,8 @@ import type { DiagnosisContextCoverage } from '~~/types/api';
 import { condenseErrorText, maskVolatile, stripAnsi } from '#shared/error-fingerprint';
 import { DIAGNOSIS_SECTIONS } from '#shared/diagnosis-sections';
 import { durationStats } from '#shared/utils/stats';
-import { computeRegressionContext, normalizeGitUrl } from './regression-context';
+import { computeRegressionContext } from './regression-context';
+import { normalizeGitUrl } from './scm/git-url';
 import { inlineCasePayloads } from './case-payloads';
 import { createScmProvider, detectScmProvider } from './scm';
 import { MAX_RAW_DIFF_BYTES } from './scm/ScmProvider';
