+++ apps/application/server/utils/compute-regression-signals.ts
@@ -2,7 +2,8 @@ import { eq } from 'drizzle-orm';
 import { projects, testRuns, testRunsCases } from '../database/schema';
 import type { DbClient as DB } from '../database';
 import { resolveRunBranch } from './run-branch';
-import { resolveDefaultBranch, FALLBACK_DEFAULT_BRANCH } from './scm/default-branch';
+import { resolveDefaultBranch } from './scm/default-branch';
+import { FALLBACK_DEFAULT_BRANCH } from './scm/git-url';
 import { selectBaselineRun } from './branch-baseline';
 
 /**
