+++ apps/application/server/utils/scm/ownership.ts
@@ -18,7 +18,7 @@
 import { eq, desc } from 'drizzle-orm';
 import { testRuns } from '../../database/schema';
 import { createScmProvider } from './index';
-import { normalizeGitUrl } from '../regression-context';
+import { normalizeGitUrl } from './git-url';
 import { TtlCache } from './cache';
 import { primaryOwnerForPath, type CompiledCodeowners } from '@piwitests/core/codeowners';
 import type { RunMetadata } from '../run-json-types';
